edk2: Change License.txt from 2-Clause BSD to BSD+Patent
Change License.txt in the root of the edk2 repository from the
2-Clause BSD License to the BSD+Patent License.

The text difference between these licenses in the patch is larger
than expected due to different choices for the position of the line
breaks.  The text from the BSD 2-Clause license and its disclaimer
are the same.  New clauses from the BSD+Patent license have been
added.

https://bugzilla.tianocore.org/show_bug.cgi?id=1373

This change is based on the following emails:

  https://lists.01.org/pipermail/edk2-devel/2019-February/036260.html
  https://lists.01.org/pipermail/edk2-devel/2018-October/030385.html

RFCs with detailed process for the license change:

  V3: https://lists.01.org/pipermail/edk2-devel/2019-March/038116.html
  V2: https://lists.01.org/pipermail/edk2-devel/2019-March/037669.html
  V1: https://lists.01.org/pipermail/edk2-devel/2019-March/037500.html

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>
Reviewed-by: Laszlo Ersek <lersek@redhat.com>