StdLib/LibC/gdtoa: Remove source file with unused ldtoa() function.
REF: https://bugzilla.tianocore.org/show_bug.cgi?id=3549

Removing the ldtoa.c file from the INF removes the ldtoa()
function from this library.  The ldtoa() function is not used
and when /WHOLEARCHIVE linker flag is enabled this unused
function makes references to other functions that are not
implemented and that generates a link failure.  Removing the
unused source file resolves the link failure.

Cc: Rebecca Cran <rebecca@nuviainc.com>
Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>
Reviewed-by: Rebecca Cran <rebecca@nuviainc.com>