Fix file descriptor leak in rename()
rename() call leads to a function that makes open() for source file to
be renamed. The resulting file descriptor is never closed. If you have
to rename a couple of files this will quickly exhaust the descriptor table.

The fix is trivial - just close the fd before returning from the function.