edk2-libc/StdLib: file descriptor leak in rename()
REF: https://bugzilla.tianocore.org/show_bug.cgi?id=4529

rename() call leads to a function that makes open() for source file
to be renamed. The resulting file descriptor is never closed.
If you have to rename a couple of files this will quickly exhaust
the descriptor table.
The fix is trivial - just close the fd before returning from the function.

Cc: Rebecca Cran <rebecca@bsdio.com>
Cc: Michael D Kinney <michael.d.kinney@intel.com>
Cc: Jayaprakash N <n.jayaprakash@intel.com>
Signed-off-by: Kloper Dimitry <dimitry.kloper@intel.com>
Reviewed-by: Michael D Kinney <michael.d.kinneyu@intel.com>
Reviewed-by: Jayaprakash N <n.jayaprakash@intel.com>