edk2-libc : make python368.inf compile conditionally in AppPkg.dsc
REF: https://bugzilla.tianocore.org/show_bug.cgi?id=4789

This commit adds necessary changes to AppPkg.dsc file for
conditional compilation of python368.inf. To enable compilation
of python368.inf users need to pass an additional parameter
to the build command as -D BUILD_PYTHON368.
Also updated the Py368ReadMe.txt and the GCCCompilationBKMs.rst
to include this additional information for building python368.inf.

Cc: Rebecca Cran <rebecca@bsdio.com>
Cc: Michael D Kinney <michael.d.kinney@intel.com>
Cc: Jayaprakash N <n.jayaprakash@intel.com>
Signed-off-by: Jayaprakash N <n.jayaprakash@intel.com>
Reviewed-by: Michael D Kinney <michael.d.kinney@intel.com>