edk2-libc: Add manual dispatch for the Python UEFI build actions
REF: https://github.com/tianocore/edk2-libc/issues/89

The current build actions are setup to dispatch on push or pull requests.
Since there is not much development activity happening on the Python UEFI,
We are seeing that the build artifacts are getting expired after 90 days.
Would like to add manual dispatch mechanism to trigger the build on need
basis.

Closes #89

Signed-off-by: Jayaprakash N <n.jayaprakash@intel.com>